Custom de/serialize for PublicKey
